+++ HISTORY.rst
@@ -175,7 +175,7 @@ History
 ------------------
 
 - Fixes
-    -
+    - "test" command had the execution of a suite and a single test switched around
 - Added
     - The run method of a single "AbstractTest" is now wrapped int a try/except statement.