+++ HISTORY.rst
@@ -176,6 +176,7 @@ History
 
 - Fixes
     - "test" command had the execution of a suite and a single test switched around
+    - fixed the "get_command_output" function which was not working
 - Added
     - The run method of a single "AbstractTest" is now wrapped int a try/except statement.